refactor: LegacyMVC 삭제
