+++ app/src/main/java/com/techcourse/controller/HomeController.java
@@ -11,9 +11,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class AnnotationHomeController {
+public class HomeController {
 
-    private static final Logger log = LoggerFactory.getLogger(AnnotationHomeController.class);
+    private static final Logger log = LoggerFactory.getLogger(HomeController.class);
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public ModelAndView show(HttpServletRequest req, HttpServletResponse res) {
