+++ app/src/main/java/com/techcourse/controller/LogoutController.java
@@ -1,16 +1,31 @@
 package com.techcourse.controller;
 
+import com.techcourse.domain.User;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
+import java.util.Optional;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public class LogoutController implements Controller {
+@Controller
+public class LogoutController {
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
+    private static final Logger log = LoggerFactory.getLogger(LogoutController.class);
+
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public ModelAndView logout(HttpServletRequest req, HttpServletResponse res) {
         final HttpSession session = req.getSession();
+
+        Optional<User> user = UserSession.getUserFrom(session);
+        user.ifPresent(value -> log.info("logged out {}", value.getAccount()));
         session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
+
+        return new ModelAndView(new JspView("redirect:/"));
     }
 }
