+++ app/src/test/java/com/techcourse/controller/HomeControllerTest.java
@@ -6,16 +6,14 @@ import static org.mockito.Mockito.when;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
-import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
+import nextstep.mvc.controller.HandlerExecution;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class AnnotationLogoutControllerTest {
-
+class HomeControllerTest {
     private AnnotationHandlerMapping handlerMapping;
 
     @BeforeEach
@@ -25,19 +23,17 @@ class AnnotationLogoutControllerTest {
     }
 
     @Test
-    void logout() throws Exception {
+    void getHomePage() throws Exception {
         final HttpServletRequest request = mock(HttpServletRequest.class);
         final HttpServletResponse response = mock(HttpServletResponse.class);
-        final HttpSession session = mock(HttpSession.class);
 
-        when(request.getRequestURI()).thenReturn("/logout");
+        when(request.getRequestURI()).thenReturn("/");
         when(request.getMethod()).thenReturn("GET");
-        when(request.getSession()).thenReturn(session);
 
         final HandlerExecution handlerExecution = (HandlerExecution) handlerMapping.getHandler(request);
         final ModelAndView modelAndView = handlerExecution.handle(request, response);
 
         JspView view = (JspView) modelAndView.getView();
-        assertThat(view.getViewName()).isEqualTo("redirect:/");
+        assertThat(view.getViewName()).isEqualTo("/index.jsp");
     }
 }
