+++ learning/src/test/java/reflection/Question.java
@@ -78,10 +78,16 @@ public class Question {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Question)) return false;
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof Question)) {
+            return false;
+        }
         Question question = (Question) o;
-        return questionId == question.questionId && countOfComment == question.countOfComment && Objects.equals(writer, question.writer) && Objects.equals(title, question.title) && Objects.equals(contents, question.contents) && Objects.equals(createdDate, question.createdDate);
+        return questionId == question.questionId && countOfComment == question.countOfComment && Objects.equals(writer,
+                question.writer) && Objects.equals(title, question.title) && Objects.equals(contents, question.contents)
+                && Objects.equals(createdDate, question.createdDate);
     }
 
     @Override
