+++ learning/src/test/java/reflection/ReflectionTest.java
@@ -76,11 +76,11 @@ class ReflectionTest {
 
         final Constructor<?> firstConstructor = questionClass.getConstructor(String.class, String.class, String.class);
         final Constructor<?> secondConstructor = questionClass.getConstructor(Long.TYPE,
-        String.class, String.class, String.class, Date.class, Integer.TYPE);
-
+                String.class, String.class, String.class, Date.class, Integer.TYPE);
 
         final Question firstQuestion = (Question) firstConstructor.newInstance("gugu", "제목1", "내용1");
-        final Question secondQuestion = (Question) secondConstructor.newInstance(0L, "gugu", "제목2", "내용2", Date.from(Instant.EPOCH), 0);
+        final Question secondQuestion = (Question) secondConstructor.newInstance(0L, "gugu", "제목2", "내용2",
+                Date.from(Instant.EPOCH), 0);
 
         assertThat(firstQuestion.getWriter()).isEqualTo("gugu");
         assertThat(firstQuestion.getTitle()).isEqualTo("제목1");
