+++ learning/src/test/java/reflection/ReflectionsTest.java
@@ -3,7 +3,6 @@ package reflection;
 import annotation.Controller;
 import annotation.Repository;
 import annotation.Service;
-import java.awt.ContainerOrderFocusTraversalPolicy;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
 import org.reflections.Reflections;
