--- mvc/src/main/java/nextstep/mvc/ControllerHandlerAdapter.java
@@ -1,23 +1,0 @@
-package nextstep.mvc;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-    @Override
-    public boolean supports(Object handler) {
-        return (handler instanceof Controller);
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
-            throws Exception {
-
-        String viewName = ((Controller) handler).execute(request, response);
-
-        return new ModelAndView(new JspView(viewName));
-    }
-}
