--- mvc/src/main/java/nextstep/mvc/controller/asis/ForwardController.java
@@ -1,20 +1,0 @@
-package nextstep.mvc.controller.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.util.Objects;
-
-public class ForwardController implements Controller {
-
-    private final String path;
-
-    public ForwardController(String path) {
-        this.path = Objects.requireNonNull(path);
-    }
-
-    @Override
-    public String execute(HttpServletRequest request, HttpServletResponse response) {
-        return path;
-    }
-}
