+++ mvc/src/main/java/nextstep/mvc/view/View.java
@@ -2,7 +2,6 @@ package nextstep.mvc.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Map;
 
 public interface View {
