+++ mvc/src/main/java/nextstep/web/NextstepServletContainerInitializer.java
@@ -4,7 +4,6 @@ import jakarta.servlet.ServletContainerInitializer;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.HandlesTypes;
-
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
