+++ mvc/src/main/java/nextstep/web/annotation/RequestMapping.java
@@ -1,11 +1,10 @@
 package nextstep.web.annotation;
 
-import nextstep.web.support.RequestMethod;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import nextstep.web.support.RequestMethod;
 
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
