+++ mvc/src/test/java/nextstep/mvc/controller/AnnotationHandlerMappingTest.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package nextstep.mvc.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -6,53 +6,47 @@ import static org.mockito.Mockito.when;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
-import nextstep.mvc.controller.tobe.HandlerExecution;
-import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class AnnotationRegisterControllerTest {
+class AnnotationHandlerMappingTest {
 
     private AnnotationHandlerMapping handlerMapping;
 
     @BeforeEach
     void setUp() {
-        handlerMapping = new AnnotationHandlerMapping("com.techcourse.controller");
+        handlerMapping = new AnnotationHandlerMapping("samples");
         handlerMapping.initialize();
     }
 
     @Test
-    void getRegisterPage() throws Exception {
+    void get() throws Exception {
         final HttpServletRequest request = mock(HttpServletRequest.class);
         final HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getRequestURI()).thenReturn("/register");
+        when(request.getAttribute("id")).thenReturn("gugu");
+        when(request.getRequestURI()).thenReturn("/get-test");
         when(request.getMethod()).thenReturn("GET");
 
         final HandlerExecution handlerExecution = (HandlerExecution) handlerMapping.getHandler(request);
         final ModelAndView modelAndView = handlerExecution.handle(request, response);
 
-        JspView view = (JspView) modelAndView.getView();
-        assertThat(view.getViewName()).isEqualTo("/register.jsp");
+        assertThat(modelAndView.getObject("id")).isEqualTo("gugu");
     }
 
     @Test
-    void register() throws Exception {
+    void post() throws Exception {
         final HttpServletRequest request = mock(HttpServletRequest.class);
         final HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getRequestURI()).thenReturn("/register");
+        when(request.getAttribute("id")).thenReturn("gugu");
+        when(request.getRequestURI()).thenReturn("/post-test");
         when(request.getMethod()).thenReturn("POST");
-        when(request.getParameter("account")).thenReturn("account");
-        when(request.getParameter("password")).thenReturn("password");
-        when(request.getParameter("email")).thenReturn("email@email.com");
 
         final HandlerExecution handlerExecution = (HandlerExecution) handlerMapping.getHandler(request);
         final ModelAndView modelAndView = handlerExecution.handle(request, response);
 
-        JspView view = (JspView) modelAndView.getView();
-        assertThat(view.getViewName()).isEqualTo("redirect:/index.jsp");
+        assertThat(modelAndView.getObject("id")).isEqualTo("gugu");
     }
 }
