--- mvc/src/test/java/samples/ManualTestController.java
@@ -1,12 +1,0 @@
-package samples;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-
-public class ManualTestController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) {
-        return null;
-    }
-}
